use latest solidity-kit (fix burn)

+++ contracts/src/solidity-kit/solc_0_8/utils/Math.sol
@@ -3,7 +3,7 @@
 pragma solidity ^0.8.0;
 
 library Math {
-    function abs(int32 x) internal pure returns (int32) {
-        return x >= 0 ? x : -x;
+    function abs(int32 x) internal pure returns (uint32) {
+        return x >= 0 ? uint32(x) : uint32(-x);
     }
 }
